refactor: move to src/actions/ folder
